requirements.txt
streamlit
python-dotenv
requests
PyPDF2
streamlit-pdf-viewer